Use the attributes of CommandContext to track fired events
