+++ modules/activiti-engine/src/main/java/org/activiti/engine/impl/bpmn/behavior/EventBasedGatewayActivityBehavior.java
@@ -13,70 +13,42 @@
 
 package org.activiti.engine.impl.bpmn.behavior;
 
-import org.activiti.bpmn.model.EventDefinition;
-import org.activiti.bpmn.model.IntermediateCatchEvent;
-import org.activiti.bpmn.model.SignalEventDefinition;
+import java.util.List;
+
 import org.activiti.engine.impl.context.Context;
+import org.activiti.engine.impl.persistence.entity.EventSubscriptionEntity;
 import org.activiti.engine.impl.persistence.entity.ExecutionEntity;
 import org.activiti.engine.impl.pvm.delegate.ActivityExecution;
-import org.activiti.engine.impl.pvm.process.ActivityImpl;
 
 
 /**
  * @author Daniel Meyer
  */
 public class EventBasedGatewayActivityBehavior extends FlowNodeActivityBehavior {
+
   
   @Override
   public void execute(ActivityExecution execution) throws Exception {
-    
-    // Continue with signal catch event activities for signal events 
-    // already fired and registered in transaction scope
-    ExecutionEntity executionEntity = (ExecutionEntity) execution;
-    ActivityImpl executionActivity = executionEntity.getActivity();
-    
-    for(ActivityImpl mayBeCatchEventActivity: executionActivity.getActivities()) {
-      if(mayBeCatchEventActivity.getActivityBehavior() instanceof IntermediateCatchEventActivityBehavior) {
-        IntermediateCatchEventActivityBehavior catchEventActivityBehavior = (IntermediateCatchEventActivityBehavior) mayBeCatchEventActivity.getActivityBehavior();
+      
+      // Continue with signal catch event activities for signal events
+      // already fired and registered in process instance execution scope
+      List<EventSubscriptionEntity> subscriptions = ((ExecutionEntity)execution).getEventSubscriptions();
 
-        SignalEventDefinition signalEventDef = findSignalEventDefinition(catchEventActivityBehavior);
-        
-        // Find matching registered signal event
-        if(signalEventDef != null) {
-          if(Context.getCommandContext().getSignalEventSessionManager()
-              .hasThrowSignalEventForExecution(execution, signalEventDef.getSignalRef())) 
+      for(EventSubscriptionEntity subscription: subscriptions) {
+          // If event already fired matches existing subscription
+          if(Context.getCommandContext().getAttribute(subscription.getEventName()) != null)
           {
-              // Switch execution context to signal catch event activity
-              executionEntity.setActivity(mayBeCatchEventActivity); 
-              
-              // Emit signal
-              catchEventActivityBehavior.signal(execution, signalEventDef.getSignalRef(), null);
-              
-              return;
+              // Execute signal event behavior 
+              subscription.eventReceived(null, false);
+
+              // Interrupt loop
+              break;
           }
-        }
       }
-    }
-    
-    // Otherwise
-    // the event based gateway doesn't really do anything
-    // ignoring outgoing sequence flows (they're only parsed for the diagram)  
+      
+      // Otherwise
+      // the event based gateway doesn't really do anything
+      // ignoring outgoing sequence flows (they're only parsed for the diagram)
   }
-  
-  protected SignalEventDefinition findSignalEventDefinition(IntermediateCatchEventActivityBehavior catchEventActivityBehavior) {
-    SignalEventDefinition signalEventDefinition = null;
-    IntermediateCatchEvent catchEvent = catchEventActivityBehavior.intermediateCatchEvent;
-    
-    if(catchEvent != null) {
-      for(EventDefinition eventDef : catchEvent.getEventDefinitions()) {
-        if(eventDef instanceof SignalEventDefinition) {
-          signalEventDefinition = (SignalEventDefinition) eventDef;
-        }
-      }
-    }
-    
-    return signalEventDefinition;
-    
-  }    
   
 }
