+++ modules/activiti-engine/src/main/java/org/activiti/engine/impl/bpmn/behavior/IntermediateCatchEventActivityBehavior.java
@@ -14,39 +14,26 @@ package org.activiti.engine.impl.bpmn.behavior;
 
 import java.util.List;
 
-import org.activiti.bpmn.model.EventDefinition;
-import org.activiti.bpmn.model.IntermediateCatchEvent;
-import org.activiti.bpmn.model.SignalEventDefinition;
 import org.activiti.engine.impl.context.Context;
+import org.activiti.engine.impl.persistence.entity.EventSubscriptionEntity;
+import org.activiti.engine.impl.persistence.entity.ExecutionEntity;
 import org.activiti.engine.impl.pvm.delegate.ActivityExecution;
 
 
 public class IntermediateCatchEventActivityBehavior extends AbstractBpmnActivityBehavior {
 
-  protected final IntermediateCatchEvent intermediateCatchEvent;
-  
-  public IntermediateCatchEventActivityBehavior(IntermediateCatchEvent intermediateCatchEvent) {
-    this.intermediateCatchEvent = intermediateCatchEvent;
-  }  
-
   @Override
   public void execute(ActivityExecution execution) throws Exception {
     
-    // Get all associated event definitions     
-    List<EventDefinition> events = intermediateCatchEvent.getEventDefinitions();
-
-    // Find signal event definition
-    for(EventDefinition event : events) {
-      if(event instanceof SignalEventDefinition) {
-        SignalEventDefinition signalEventDefinition = (SignalEventDefinition) event;
-        
-        // Check already published signal events registered in the transaction context   
-        if(Context.getCommandContext().getSignalEventSessionManager()
-            .hasThrowSignalEventForExecution(execution, signalEventDefinition.getSignalRef()))
-        {
-            // Leave execution if matching throw signal found 
-           leave(execution);
-        }
+    // Find active subscriptions
+    List<EventSubscriptionEntity> subscriptions = ((ExecutionEntity)execution).getEventSubscriptions();
+    
+    for(EventSubscriptionEntity subscription: subscriptions) {
+      // Check already published signal events registered in the transaction context   
+      if(Context.getCommandContext().getAttribute(subscription.getEventName()) != null)
+      {
+        // Signal event for matching throw signal 
+        subscription.eventReceived(null, false);
       }
     }
