+++ modules/activiti-engine/src/main/java/org/activiti/engine/impl/interceptor/CommandContext.java
@@ -52,7 +52,6 @@ import org.activiti.engine.impl.persistence.entity.ProcessDefinitionEntityManage
 import org.activiti.engine.impl.persistence.entity.ProcessDefinitionInfoEntityManager;
 import org.activiti.engine.impl.persistence.entity.PropertyEntityManager;
 import org.activiti.engine.impl.persistence.entity.ResourceEntityManager;
-import org.activiti.engine.impl.persistence.entity.ThrowSignalEventSessionManager;
 import org.activiti.engine.impl.persistence.entity.TableDataManager;
 import org.activiti.engine.impl.persistence.entity.TaskEntityManager;
 import org.activiti.engine.impl.persistence.entity.UserIdentityManager;
@@ -377,10 +376,6 @@ public class CommandContext {
     return getSession(EventSubscriptionEntityManager.class);
   }
   
-  public ThrowSignalEventSessionManager getSignalEventSessionManager() {
-    return getSession(ThrowSignalEventSessionManager.class); 
-  }  
-
   public Map<Class< ? >, SessionFactory> getSessionFactories() {
     return sessionFactories;
   }
