+++ modules/activiti-engine/src/test/resources/org/activiti/engine/test/bpmn/event/signal/SignalEventTest.testSignalRaceConditionsAfterParallelGatewayGlobal.bpmn20.xml
