+++ modules/activiti-engine/src/test/resources/org/activiti/engine/test/bpmn/event/signal/SignalEventTest.testSignalThrowCatchEventGateway.bpmn20.xml
